Fix #359 remove checkout templates from theme & move checkout to own blueprint
